refactore books array
