+++ src/utils/randomId.js
@@ -1,0 +1,3 @@
+const randomId = () => Math.floor(Math.random() * 100);
+
+export default randomId;
